handlers of user and admin plus UserMiddleware

+++ internal/usecase/admin/admin.go
@@ -2,7 +2,7 @@ package admin
 
 import (
 	"ads-service/internal/domain/entities"
-	usecaseerr "ads-service/internal/errs/usecaseerr"
+	"ads-service/internal/errs/usecaseerr"
 	"context"
 	"log"
 	"time"
@@ -22,8 +22,14 @@ func (s *service) GetAllAds(ctx context.Context) ([]entities.Ad, error) {
 }
 
 func (s *service) DeleteAd(ctx context.Context, adID int) error {
+	if adID <= 0 {
+		log.Println("invalid ad ID")
+		return usecaseerr.ErrInvalidParams
+	}
+
 	err := s.adRepo.Delete(ctx, adID)
 	if err != nil {
+		log.Println("error deleting ad:", err)
 		return usecaseerr.ErrDeletingAd
 	}
